Sửa lỗi build CS1503 trong ControllerSession.cs
Lỗi CS1503 (uint -> int) do frame.Width/frame.Height từ VpxVideoEncoder.DecodeVideo()
trả về kiểu uint, trong khi I420ToBitmap() và RemoteVideoForm.UpdateFrame() khai báo
nhận int -> thêm ép kiểu (int) tường minh. Log lỗi thật từ CI (windows-latest, .NET 8):

ControllerSession.cs(104,54): error CS1503: Argument 2: cannot convert from 'uint' to 'int'
ControllerSession.cs(104,67): error CS1503: Argument 3: cannot convert from 'uint' to 'int'
ControllerSession.cs(105,41): error CS1503: Argument 2: cannot convert from 'uint' to 'int'
ControllerSession.cs(105,54): error CS1503: Argument 3: cannot convert from 'uint' to 'int'

+++ windows-agent/ControllerSession.cs
@@ -101,8 +101,8 @@ public sealed class ControllerSession : IAsyncDisposable
             var decodedFrames = _codec.DecodeVideo(payload, VideoPixelFormatsEnum.I420, VideoCodecsEnum.VP8);
             foreach (var frame in decodedFrames)
             {
-                var bmp = I420ToBitmap(frame.Sample, frame.Width, frame.Height);
-                _form?.UpdateFrame(bmp, frame.Width, frame.Height);
+                var bmp = I420ToBitmap(frame.Sample, (int)frame.Width, (int)frame.Height);
+                _form?.UpdateFrame(bmp, (int)frame.Width, (int)frame.Height);
             }
         }
         catch (Exception ex)
